refactor: split different subcmd
