+++ nyaruko/config.py
@@ -1,0 +1,4 @@
+import configparser
+
+config = configparser.ConfigParser()
+config.read('config.ini')
