+++ nyaruko/main.py
@@ -1,24 +1,11 @@
 """
 Nyaruko is a command line tool for nyaruko.
 """
-import os
-import configparser
 import click
-from telegram import Update
-from telegram.ext import ApplicationBuilder, CommandHandler, ContextTypes
-
-config = configparser.ConfigParser()
-config.read('config.ini')
-
-
-async def telegram_start(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
-    """
-    Handles the /start command.
-    """
-    await update.message.reply_text(
-        f"Hi! {update.message.from_user.first_name}. I'm Nyaruko, your personal assistant. " +
-        "I'm still under development, but you can talk to me if you want."
-    )
+from nyaruko.command.ascii import show_ascii
+from nyaruko.command.bot import run_bot
+from nyaruko.command.sync import sync
+from nyaruko.command.list import list_article
 
 
 @click.group()
@@ -32,46 +19,7 @@ def cli(ctx, debug):
     ctx.obj['DEBUG'] = debug
 
 
-@cli.command("list")
-def list_article():
-    """
-    Lists all article from the obsidian.
-    """
-    article_dir = config["default"]["article_dir"]
-
-    for root, dirs, files in os.walk(article_dir):
-        for file in files:
-            if file.endswith(".md"):
-                article_path = os.path.join(root, file)[len(article_dir) + 1:]
-                click.echo(article_path)
-
-
-@cli.command("ascii")
-@click.option("--big", is_flag=True, help="Print the big ascii art of nyaruko.")
-def show_ascii(big):
-    """
-    Prints the ASCII art of nyaruko.
-    """
-    art_path = "data/ascii-art-big.txt" if big else "data/ascii-art.txt"
-    with open(os.path.join(os.path.dirname(__file__), art_path), "r", encoding="utf-8") as file:
-        click.echo(file.read())
-
-
-@cli.command("bot")
-@click.option("--daemon", is_flag=True, help="Run the bot as a daemon.")
-def run_bot(daemon):
-    """
-    Runs the telegram bot.
-    """
-    if not daemon:
-        click.echo(
-            "Hi! I'm Nyaruko, your personal assistant. "
-            "I'm still under development, but you can talk to me if you want."
-        )
-        return None
-
-    app = ApplicationBuilder().get_updates_proxy_url(config["telegram"]["proxy"]).token(
-        config["telegram"]["token"]).build()
-
-    app.add_handler(CommandHandler("start", telegram_start))
-    app.run_polling()
+cli.add_command(sync)
+cli.add_command(show_ascii, "ascii")
+cli.add_command(list_article, "list")
+cli.add_command(run_bot, "bot")
